requirements.txt
groq>=0.4.0
langchain>=0.3.0
langchain-groq>=0.1.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
httpx>=0.27.0
datasets>=2.19.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
requests>=2.31.0
